requirements.txt

streamlit
pdfplumber
langchain
faiss-cpu
openai
python-dotenv
sentence-transformers
